otmp-project/otmp-service/otmp-masterdata/src/test/java/com/microservices/sdms/masterdata/service/impl/BizBaseGtnCategoryServiceImplTest.java
/*
package com.microservices.otmp.masterdata.service.impl;

import com.microservices.otmp.masterdata.mapper.BizBaseGtnCategoryMapper;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.MockitoJUnitRunner;

import java.util.Arrays;
import java.util.Collections;

@RunWith(MockitoJUnitRunner.class)
public class BizBaseGtnCategoryServiceImplTest {
    @Mock
    private BizBaseGtnCategoryMapper mockBizBaseGtnCategoryMapper;

    @InjectMocks
    private BizBaseGtnCategoryServiceImpl bizBaseGtnCategoryServiceImplUnderTest;

    @Test
    public void testSelectBizBaseGtnCategoryByRegion() {
        // Setup
        // Configure BizBaseGtnCategoryMapper.selectBizBaseGtnCategoryByRegionMarketCode(...).
        final BizBaseGtnCategory bizBaseGtnCategory = new BizBaseGtnCategory();
        bizBaseGtnCategory.setGtnType("gtnType");
        bizBaseGtnCategory.setGtnCategory("gtnCategory");
        bizBaseGtnCategory.setStatus("status");
        when(mockBizBaseGtnCategoryMapper.selectBizBaseGtnCategoryByRegionMarketCode("region")).thenReturn(bizBaseGtnCategory);

        // Run the test
        final BizBaseGtnCategory result = bizBaseGtnCategoryServiceImplUnderTest.selectBizBaseGtnCategoryByRegion("region");

        assertEquals(bizBaseGtnCategory, result);
        // Verify the results
    }

    @Test
    public void testSelectBizBaseGtnCategoryList() {
        // Setup
        final BizBaseGtnCategory bizBaseGtnCategory = new BizBaseGtnCategory();
        bizBaseGtnCategory.setGtnType("gtnType");
        bizBaseGtnCategory.setGtnCategory("gtnCategory");
        bizBaseGtnCategory.setStatus("status");

        // Configure BizBaseGtnCategoryMapper.selectBizBaseGtnCategoryList(...).
        final BizBaseGtnCategory bizBaseGtnCategory1 = new BizBaseGtnCategory();
        bizBaseGtnCategory1.setGtnType("gtnType");
        bizBaseGtnCategory1.setGtnCategory("gtnCategory");
        bizBaseGtnCategory1.setStatus("status");
        final List<BizBaseGtnCategory> bizBaseGtnCategories = Arrays.asList(bizBaseGtnCategory1);
        when(mockBizBaseGtnCategoryMapper.selectBizBaseGtnCategoryList(any(BizBaseGtnCategory.class))).thenReturn(bizBaseGtnCategories);

        // Run the test
        final List<BizBaseGtnCategory> result = bizBaseGtnCategoryServiceImplUnderTest.selectBizBaseGtnCategoryList(bizBaseGtnCategory);
        assertEquals(bizBaseGtnCategories, result);
        // Verify the results
    }

    @Test
    public void testSelectBizBaseGtnCategoryList_BizBaseGtnCategoryMapperReturnsNoItems() {
        // Setup
        final BizBaseGtnCategory bizBaseGtnCategory = new BizBaseGtnCategory();
        bizBaseGtnCategory.setGtnType("gtnType");
        bizBaseGtnCategory.setGtnCategory("gtnCategory");
        bizBaseGtnCategory.setStatus("status");

        when(mockBizBaseGtnCategoryMapper.selectBizBaseGtnCategoryList(any(BizBaseGtnCategory.class))).thenReturn(Collections.emptyList());

        // Run the test
        final List<BizBaseGtnCategory> result = bizBaseGtnCategoryServiceImplUnderTest.selectBizBaseGtnCategoryList(bizBaseGtnCategory);

        // Verify the results
        assertEquals(Collections.emptyList(), result);
    }

    @Test
    public void testInsertBizBaseGtnCategory() {
        // Setup
        final BizBaseGtnCategory bizBaseGtnCategory = new BizBaseGtnCategory();
        bizBaseGtnCategory.setGtnType("gtnType");
        bizBaseGtnCategory.setGtnCategory("gtnCategory");
        bizBaseGtnCategory.setStatus("status");

        when(mockBizBaseGtnCategoryMapper.insertBizBaseGtnCategory(any(BizBaseGtnCategory.class))).thenReturn(0);

        // Run the test
        final int result = bizBaseGtnCategoryServiceImplUnderTest.insertBizBaseGtnCategory(bizBaseGtnCategory);

        // Verify the results
        assertEquals(0, result);
    }

    @Test
    public void testUpdateBizBaseGtnCategory() {
        // Setup
        final BizBaseGtnCategory bizBaseGtnCategory = new BizBaseGtnCategory();
        bizBaseGtnCategory.setGtnType("gtnType");
        bizBaseGtnCategory.setGtnCategory("gtnCategory");
        bizBaseGtnCategory.setStatus("status");

        when(mockBizBaseGtnCategoryMapper.updateBizBaseGtnCategory(any(BizBaseGtnCategory.class))).thenReturn(0);

        // Run the test
        final int result = bizBaseGtnCategoryServiceImplUnderTest.updateBizBaseGtnCategory(bizBaseGtnCategory);

        // Verify the results
        assertEquals(0, result);
    }

    @Test
    public void testDeleteBizBaseGtnCategoryByRegions() {
        // Setup
        when(mockBizBaseGtnCategoryMapper.deleteBizBaseGtnCategoryByRegionMarketCodes(any(String[].class))).thenReturn(0);

        // Run the test
        final int result = bizBaseGtnCategoryServiceImplUnderTest.deleteBizBaseGtnCategoryByRegions(new String[]{"value"});

        // Verify the results
        assertEquals(0, result);
    }

    @Test
    public void testDeleteBizBaseGtnCategoryByRegion() {
        // Setup
        when(mockBizBaseGtnCategoryMapper.deleteBizBaseGtnCategoryByRegionMarketCode("region")).thenReturn(0);

        // Run the test
        final int result = bizBaseGtnCategoryServiceImplUnderTest.deleteBizBaseGtnCategoryByRegion("region");

        // Verify the results
        assertEquals(0, result);
    }
}
*/
